python/pyproject.toml
[project]
name = "geoengine"
version = "0.8.0"
authors = [{ name = "Geo Engine GmbH", email = "info@geoengine.de" }]
description = "A package for easy access to Geo Engine instances"
classifiers = [
    "Programming Language :: Python :: 3",
    "Topic :: Database :: Front-Ends",
    "Topic :: Scientific/Engineering :: GIS",
    "Topic :: Software Development :: Libraries",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Information Technology",
]
readme = { file = "README.md", content-type = "text/markdown" }
license-files = ["LICENSE"]
requires-python = ">=3.10"
dependencies = [
    "geoengine-openapi-client == 0.0.33",
    "geopandas >=1.0,<2.0",
    "matplotlib >=3.6,<3.11",
    "numpy >=1.23,<2.5",
    "owslib >=0.27,<0.36",
    "pillow >=10.0,<13",
    "pyarrow >=17.0,<24",
    "python-dotenv >=0.19,<1.3",
    "rasterio >=1.3,<2",
    "requests >= 2.26,<3",
    "rioxarray >=0.9.1, <0.23",
    "StrEnum >=0.4.6,<0.5",               # TODO: use from stdlib when `python_requires = >=3.11`
    "vega >= 3.5,<4.2",
    "websockets >= 14.2,<17",
    "xarray >=0.19,<2026.3",
    "urllib3 >= 2.1, < 2.7",
    "pydantic >= 2.10.6, < 2.13",
    'skl2onnx >=1.20,<1.21',
]

[project.urls]
Homepage = "https://www.geoengine.io"
Repository = "https://github.com/geo-engine/geoengine-python"

[project.optional-dependencies] # TODO: use [dependency-groups] in the future
dev = [
    "build >=0.7,<1.5",
    "mypy >=1.14,<2.0",
    "pdoc3 >=0.10,<0.12",
    "ruff >=0.12.3,<0.16",         # formatter & linter
    "setuptools >=62,<83",
    "twine >=3.4,<6.3",            # PyPI
    "types-requests >=2.26,<3",    # mypy type hints
    "types-setuptools >=71.1,<83", # mypy type hints
    "wheel >=0.37,<0.47",
]
test = [
    "psycopg >=3.2,<4",
    "pytest >=6.3,<10",
    "pytest-cov >=6.0,<7.1",
    "requests_mock >=1.9,<2",
    "scikit-learn >=1.7,<1.9",
]
examples = [
    "cartopy >=0.22,<0.26",    # for WMS example
    "ipympl >=0.9.4,<0.11",    # for ML example
    "ipyvuetify >=1.10,<1.12", # for ML app
    "ipywidgets >=8.1.5,<9",   # for ML example
    "nbconvert >=7,<8",        # convert notebooks to Python
    "scipy >=1.7,<2",          # for WMS example
    "voila >=0.5,<0.6",        # for ML app
]

[build-system]
requires = ["setuptools>=62"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # Pyflakes
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "I",   # isort
]

[tool.ruff.lint.per-file-ignores]
"geoengine/__init__.py" = [
    "F401", # module imported but unused
]
"geoengine/workflow_builder/__init__.py" = [
    "F401", # module imported but unused
]
"tests/__init__.py" = [
    "F401", # module imported but unused
]

[tool.pytest.ini_options]
testpaths = ["tests"]
